+++ internal/app/sdk.go
@@ -13,6 +13,7 @@ type SDK interface {
 	GetDriveItemByPath(path string) (onedrive.DriveItem, error)
 	GetDriveItemChildrenByPath(path string) (onedrive.DriveItemList, error)
 	GetDrives() (onedrive.DriveList, error)
+	GetDefaultDrive() (onedrive.Drive, error)
 	CreateFolder(parentPath string, folderName string) (onedrive.DriveItem, error)
 	DownloadFile(remotePath, localPath string) error
 	CreateUploadSession(remotePath string) (onedrive.UploadSession, error)
@@ -44,6 +45,11 @@ func (s *OneDriveSDK) GetDriveItemChildrenByPath(path string) (onedrive.DriveIte
 // GetDrives calls the real onedrive.GetDrives function.
 func (s *OneDriveSDK) GetDrives() (onedrive.DriveList, error) {
 	return onedrive.GetDrives(s.client)
+}
+
+// GetDefaultDrive calls the real onedrive.GetDefaultDrive function.
+func (s *OneDriveSDK) GetDefaultDrive() (onedrive.Drive, error) {
+	return onedrive.GetDefaultDrive(s.client)
 }
 
 // CreateFolder calls the real onedrive.CreateFolder function.